Common response class extracted to its own module. Rows data is now stored including field name, following the CARTO style. Closes #35
