+++ src/core/tests/test_data_source_base.py
@@ -2,7 +2,8 @@ import os
 from unittest import TestCase, mock
 
 from ..caches.base import LongitudeCache
-from ..data_sources.base import DataSource, DataSourceQueryConfig, LongitudeQueryResponse
+from ..data_sources.base import DataSource, DataSourceQueryConfig
+from ..common.query_response import LongitudeQueryResponse
 
 
 def load_raw_text(filename):
