+++ src/core/tests/test_data_source_carto.py
@@ -54,7 +54,7 @@ class TestCartoDataSource(TestCase):
     def test_succesful_query(self):
         ds = CartoDataSource()
         ds._sql_client = mock.MagicMock()
-        ds._sql_client.send.return_value = {'rows': [], 'time': 42.0, 'fields': {}}
+        ds._sql_client.send.return_value = {'rows': [], 'time': 42.0, 'fields': {}, 'total_rows': 0}
         result = ds.query('some query')
         ds._sql_client.send.assert_called_with('some query', do_post=False, format='json', parse_json=True)
         self.assertEqual([], result.rows)
